Created NamedItemHolder class
Class essentially has same functionality as the former ItemHolder class, but now inherits a more primitive class (the new ItemHolder class) which merely holds items and performs various functions relating to the items it holds. In comparison, the NamedItemHolder class contains all of the functions related to ItemHolder objects that have names.