Changed type of purchasedItems PIV to ItemHolder
Changed from a StockItem vector. This way, the purchasedItems instance variable has the functionality of checking to see if a passed StockItem already exists in the list of purchased items, and adds to the quantity of the existing instance rather than adding a new object. Fixes #1.

Also changed all instances of methods called on purchasedItems to reflect the ItemHolder equivalent.